# site/api/openapi.yaml
openapi: 3.0.3
info:
  title: MediaNest API
  description: |
    MediaNest API provides comprehensive media management, Plex integration, YouTube downloading, and system administration capabilities.

    ## Authentication
    - **Primary**: Plex OAuth authentication
    - **Secondary**: Password-based authentication for admin users
    - **Token**: JWT tokens stored in HTTP-only cookies
    - **CSRF**: Required for state-changing operations

    ## Features
    - Media search and request management
    - Plex server integration and library browsing
    - YouTube video downloading with progress tracking
    - Real-time WebSocket events
    - Comprehensive admin panel
    - Performance monitoring and optimization

  version: 1.0.0
  contact:
    name: MediaNest API Support
  license:
    name: MIT
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: https://api.medianest.app/api/v1
    description: Production server
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /auth/plex/pin:
    post:
      tags:
        - Authentication
      summary: Generate Plex PIN for OAuth
      description: Creates a PIN that users can authorize on plex.tv/link
      security: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                clientName:
                  type: string
                  default: 'MediaNest'
                  example: 'MediaNest'
      responses:
        '200':
          description: PIN generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlexPinResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /auth/session:
    get:
      tags:
        - Authentication
      summary: Get current session
      description: Returns information about the current authenticated user
      responses:
        '200':
          description: Session information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Ends the current user session
      parameters:
        - $ref: '#/components/parameters/CSRFToken'
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Media Management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
  parameters:
...
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 'Operation completed successfully'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: 'VALIDATION_ERROR'
            message:
              type: string
              example: 'Request validation failed'
            details:
              type: string
              example: 'Additional error details'
        correlationId:
          type: string
          example: 'abc123-def456'
...
    PlexPinResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: string
              example: '12345'
            code:
              type: string
              example: 'ABCD1234'
            qrUrl:
              type: string
              format: uri
              example: 'https://plex.tv/link/?pin=ABCD1234'
            expiresIn:
              type: integer
              example: 900
...
    SessionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: string
          example: 'user-123'
        username:
          type: string
          example: 'plex_user'
        email:
          type: string
          format: email
          example: 'user@example.com'
        role:
          type: string
          enum: [admin, user]
          example: 'user'
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
          nullable: true

    # Media Schemas
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'UNAUTHORIZED'
              message: 'Authentication required'
...
    ServiceUnavailable:
      description: Service temporarily unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 'SERVICE_UNAVAILABLE'
              message: 'Service temporarily unavailable'
...